api-gateway/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
babel==2.17.0
backrefs==5.9
bcrypt==5.0.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.0
ghp-import==2.1.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.23
mkdocs-material-extensions==1.3.1
mypy_extensions==1.1.0
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
pyyaml_env_tag==1.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
